pulumi/handlers/docker-summarize/requirements.txt
openai
pymongo
newspaper3k
python-dotenv
newsapi-python